# Support CIGARs longer than 65535 operations