--- a/src/cljam/io/bam/encoder.clj
+++ b/src/cljam/io/bam/encoder.clj
@@ -93,8 +93,20 @@
        read-length
        ^long (get-options-size aln))))
 
+(defn- add-cigar-to-options
+  [options cigar]
+  (cons
+   {:CG {:type "B",:value (str "I," (cstr/join "," (cigar/encode-cigar cigar)))}}
+   options))
+
 (defn encode-alignment [wrtr aln refs]
-  (let [aln (update aln :seq #(if (= % "*") "" %))]
+  (let [aln (update aln :seq #(if (= % "*") "" %))
+        cigar-ops-count (cigar/count-op (:cigar aln))
+        [encoded-cigar cigar-ops-count opts*]
+        (if (> cigar-ops-count 65535)
+          [(cigar/->placeholder (:cigar aln))
+           2 (add-cigar-to-options (:options aln) (:cigar aln))]
+          [(cigar/encode-cigar (:cigar aln)) cigar-ops-count (:options aln)])]
     ;; refID
     (lsb/write-int wrtr (or (refs/ref-id refs (:rname aln)) -1))
     ;; pos
@@ -104,7 +116,7 @@
     (lsb/write-ubyte wrtr (short (:mapq aln)))
     (lsb/write-ushort wrtr (sam-util/compute-bin aln))
     ;; flag_nc
-    (lsb/write-ushort wrtr (cigar/count-op (:cigar aln)))
+    (lsb/write-ushort wrtr cigar-ops-count)
     (lsb/write-ushort wrtr (:flag aln))
     ;; l_seq
     (lsb/write-int wrtr (.length ^String (:seq aln)))
@@ -118,16 +130,14 @@
     (lsb/write-string wrtr (:qname aln))
     (lsb/write-bytes wrtr (byte-array 1 (byte 0)))
     ;; cigar
-    (doseq [cigar (cigar/encode-cigar (:cigar aln))]
-      (lsb/write-int wrtr cigar))
+    (doseq [cigar encoded-cigar] (lsb/write-int wrtr cigar))
     ;; seq
     (lsb/write-bytes wrtr (seq/str->compressed-bases (:seq aln)))
     ;; qual
     (lsb/write-bytes wrtr (encode-qual aln))
-
     ;; options
-    (doseq [op (:options aln)]
-      (let [[tag value] (first (seq op))]
+    (doseq [opt opts*]
+      (let [[tag value] (first (seq opt))]
         (lsb/write-short
          wrtr
          (short (bit-or (bit-shift-left (byte (second (name tag))) 8)