… test-resources/bam/long_cigar_operations.bam
… test-resources/sam/long_cigar_operations.sam