This PR resolves #142 .
This PR enables decoding and encoding support for SAM/BAM files containing CIGARs with operation longer than 65535 operations.

According to [Sequence Alignment/Map Format Specification](http://samtools.github.io/hts-specs/SAMv1.pdf), if the CIGAR in the BAM file alignment has an operation longer than 65535, the CIGAR value is replaced by a placeholder of the form 'kSmN' and the original CIGAR value is moved into the options.

Previously, cljam's decoding and encoding process did not support this behavior.

In particular, I would like you to pay attention to the following items in your review.
- Are function and variable names reasonable?
- Should comments be added to make the code easier to understand?
- Does this modification significantly reduce processing speed?