@matsutomo81 2022-09-28
I added benchmark scripts for BAM I/O and did a run.

**Benchmark**

https://github.com/chrovis/cljam/blob/5d1b98097197136a48171198803ed9233c72d465/bench/cljam/io/bam/encoder_bench.clj#L7-L19

https://github.com/chrovis/cljam/blob/5d1b98097197136a48171198803ed9233c72d465/bench/cljam/io/bam/encoder_bench.clj#L21-L31

https://github.com/chrovis/cljam/blob/5d1b98097197136a48171198803ed9233c72d465/bench/cljam/io/bam/decoder_bench.clj#L7-L14

https://github.com/chrovis/cljam/blob/5d1b98097197136a48171198803ed9233c72d465/bench/cljam/io/bam/decoder_bench.clj#L16-L21

master

```clojure
encode-alignment-short-bench (encoder_bench.clj:7)

  ["test-resources/sam/test.sam"]
  time: 100.840645 µs, sd: 0.006130 ns

  ["test-resources/sam/medium.sam"]
  time: 281.704315 ms, sd: 32.336337 µs

  ["test-resources/sam/codec_benchmark.sam"]
  time: 1.128733 sec, sd: 119.049867 µs

encode-alignment-long-bench (encoder_bench.clj:21)

  ["/Users/tmatsuda/clj/cljam-master/cljam/.cavia/large.bam"]
  time: 25.498248 sec, sd: 34.642494 ms

decode-bam-alignment-short-bench (decoder_bench.clj:7)

  ["test-resources/bam/test.bam"]
  time: 65.088405 µs, sd: 0.001436 ns

  ["test-resources/bam/medium.bam"]
  time: 34.315355 ms, sd: 352.526299 ns

  ["test-resources/bam/codec_benchmark.bam"]
  time: 102.724572 ms, sd: 3.677748 µs

decode-bam-alignment-long-bench (decoder_bench.clj:16)

  ["/Users/tmatsuda/clj/cljam-master/cljam/.cavia/large.bam"]
  time: 23.107370 sec, sd: 543.051677 ms
```

PR

```clojure
encode-alignment-short-bench (encoder_bench.clj:7)

  ["test-resources/sam/test.sam"]
  time: 100.380541 µs, sd: 0.010300 ns

  ["test-resources/sam/medium.sam"]
  time: 281.021606 ms, sd: 13.082381 µs

  ["test-resources/sam/codec_benchmark.sam"]
  time: 1.134700 sec, sd: 100.222050 µs

encode-alignment-long-bench (encoder_bench.clj:21)

  ["/Users/tmatsuda/clj/cljam_original/cljam/.cavia/large.bam"]
  time: 25.633702 sec, sd: 91.624528 ms
  
  decode-bam-alignment-short-bench (decoder_bench.clj:7)

  ["test-resources/bam/test.bam"]
  time: 67.686084 µs, sd: 0.002714 ns

  ["test-resources/bam/medium.bam"]
  time: 35.469087 ms, sd: 352.967443 ns

  ["test-resources/bam/codec_benchmark.bam"]
  time: 106.913516 ms, sd: 3.472430 µs

decode-bam-alignment-long-bench (decoder_bench.clj:16)

  ["/Users/tmatsuda/clj/cljam_original/cljam/.cavia/large.bam"]
  time: 22.955122 sec, sd: 111.101131 ms
```