@matsutomo81 2022-09-30
Thanks for the review!
Code and benchmark results have been updated in response to the review.