@matsutomo81 2022-10-05
While editing the code, I noticed that in some of the `medium-bam-file` alignments, the cigar was blank (`:cigar ""`).
Is such a case possible in the actual data? Should we support this kind of pattern?

https://github.com/chrovis/cljam/blob/843e30b4aaf34e2fe79d070f1a2122097eb78b99/test/cljam/test_common.clj#L143