@matsutomo81 2022-11-04
I have updated the commits and 'PR' benchmark results : https://github.com/chrovis/cljam/pull/263#issuecomment-1260352904.